requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
dill==0.3.9
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==6.0.0
mccabe==0.7.0
passlib==1.7.4
platformdirs==4.3.6
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pylint==3.3.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.34.0
